+++ src/systems/states/transitions.rs
@@ -1,6 +1,7 @@
 use bevy::{
     audio::{PlaybackMode, Volume},
     prelude::*,
+    window::Cursor,
 };
 
 use crate::core::resources::assets::AudioAssets;
@@ -10,6 +11,7 @@ use super::{ForState, GameState};
 pub(crate) fn game_setup(
     mut commands: Commands,
     audios: Res<AudioAssets>,
+    mut window: Query<&mut Window>,
     mut next_state: ResMut<NextState<GameState>>,
 ) {
     commands.spawn((
@@ -23,6 +25,15 @@ pub(crate) fn game_setup(
         },
         Name::new("Ambient Music"),
     ));
+
+    // TODO: Reenable cursor when returning to StartMenu
+    //       or when entering a menu interface.
+    // TODO: Ideally also reenable cursor for debug menu
+    //       in development mode.
+    window.single_mut().cursor = Cursor {
+        visible: false,
+        ..default()
+    };
 
     next_state.set(GameState::Active);
 }